pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devops-cli"
version = "1.0.0"
description = "A powerful CLI tool for DevOps workflows - Git, Deployments, Logs, Health Checks"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Team", email = "team@startup.com"}
]
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "paramiko>=3.0.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
devops = "devops_cli.main:app"
dx = "devops_cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
